Move Capture items
