--- OAT/Captures/RuleCapture.cs
@@ -1,16 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace Microsoft.CST.OAT.Captures
-{
-    public class RuleCapture
-    {
-        public RuleCapture(Rule r, List<ClauseCapture> captures)
-        {
-            Rule = r;
-            Captures = captures;
-        }
-
-        public Rule Rule { get; }
-        public List<ClauseCapture> Captures { get; }
-    }
-}
