+++ OAT/ClauseCapture.cs
